geofence rooms updating

+++ app/src/main/java/pt/ulisboa/tecnico/cmov/conversationalist/models/GeoCage.java
@@ -5,8 +5,8 @@ import java.io.Serializable;
 public class GeoCage implements Serializable {
 
     public String chatroomID;
-    public Long latitude;
-    public Long longitude;
+    public Double latitude;
+    public Double longitude;
     public Long radius;
 
 }
